requirements.txt
fastapi
uvicorn
pydantic
starlette
alembic
psycopg2
sqlalchemy~=1.4.7
python-dotenv
SQLAlchemy-serializer